Merged PR 5: Merge signup-login working state into development
I believe I got all of the base requirements done for the signup, and all but "let the user change their password" in the login requirements (they can reset their password but there's no navigation to the account management page to change passwords while signed-in).

I'll try to get that done tomorrow if I can.